requirements.txt
# Core Mobile Testing
Appium-Python-Client==3.1.0    # Appium client for Python
selenium==4.16.0                # Selenium for WebDriver base

# Testing Framework
pytest==7.4.3                   # Test framework
pytest-xdist==3.5.0            # Parallel test execution
pytest-html==4.1.1             # HTML reports
pytest-rerunfailures==12.0     # Retry flaky tests
allure-pytest==2.13.2          # Allure reporting

# Configuration & Data
pyyaml==6.0.1                  # YAML config files
python-dotenv==1.0.0           # Environment variables
faker==22.0.0                  # Test data generation

# Utilities
loguru==0.7.2                  # Better logging
pillow==10.1.0                 # Image processing (screenshots)
opencv-python==4.9.0.80        # Image comparison
numpy==1.26.2                  # Numerical operations

# API Testing (for hybrid apps)
requests==2.31.0               # HTTP requests

# Mobile-Specific
pure-python-adb==0.3.0.dev0    # Android Debug Bridge
